Sparse matrices now admit as constructors a vector of triplets.

+++ src/sparse/mmult_sparse_tensor.h
@@ -76,7 +76,7 @@ static void mult_sp_t(elt_t *dest, const index *row_start, const index *column,
 template <typename elt_t>
 static inline Tensor<elt_t> do_mmult(const Sparse<elt_t> &m1,
                                      const Tensor<elt_t> &m2) {
-  Indices dims(static_cast<size_t>(m2.rank()));
+  Indices dims(m2.rank());
   index l_len = 1;
   for (index k = 1, N = m2.rank(); k < N; k++) {
     dims.at(k) = m2.dimension(k);
